Making sure that submitit works, and thus scheduling jobs on SLURM will work, by moving the parts to be pickled in a proper module.
